requirements.txt
torch
numpy
pyyaml
wandb
python-dotenv